requirements.txt
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.2
numpy==2.1.0
folium==0.18.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
haversine==2.8.1
streamlit_folium==0.23.2
pillow==10.4.0
DateTime==5.5